# src/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Classroom API Rest",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/users": {
            "get": {
                "description": "Get all Users",
                "tags": ["Users"],
                "responses": {
                    "200": {
                        "description": "we successfully got all users"
                    },
                    "400": {
                        "description": "we generate an error"
                    }
                }
            },
            "post": {
                "description": "Create new user",
                "tags": ["Users"],
                "responses": {
                    "200": {
                        "description": "we successfully create a new user"
                    },
                    "400": {
                        "description": "we generate an error"
                    }
                }
            }
        },
...
    }
}